Fix bug with pushing boulder left to wall
There was this bug where pushing boulders to the left, while the boulder
was hitting a wall, would continually place everything 1 pixel to the
right.
This was actually because the tiles's hitbox is 15x15, not 16x16. This
makes it so the boulder still "sees" the wall after collision detection,
which "communicates" to the player that the boulder is hitting a wall,
and should be treated solid.
Also fixes some other weirdness, like graphically running up to a wall,
on the left you'd see one pixel that on the right you didn't.